date/time - execute queries on ISO 8601 datetimes or times
This commit allows you to execute equality and comparison operator
queries on fields that adhere either to ISO 8601 formatting (`YYYY-MM-DD
[T]HH:mm:ss[Z]`, where the separator `T` and timezone `Z` are optional)
or is a time like `HH:mm:ss`.

Tables contains such conformant data do not need to be explicitly marked
as so programmatically or during the search query. tablequeryjs
aggressively attempts to decode all strings in the first data row to
determine which columns contain such fields.

Currently there is no different between LT and LTE, or GT and GTE.
Moreover, datetime queries are only precise to the nearest day.