Fix `grunt build`
This is the sledgehammer approach to fixing this, but here goes for
documentation purposes that people can find via google on a rejected
pull request if nothing else.

After doing `npm install; grunt build`[1] as per the README, I was
getting this error:

    Running "bower:install" (bower) task
    Fatal error: Arguments to path.join must be strings

A bit of googling suggests that it's a known issue with some particular
version of grunt or its dependencies. I fixed it by removing all the
version numbers from package.json and running this:

    npm install --save bower
    npm install --save-dev \
        grunt \
        grunt-contrib-qunit \
        grunt-cli \
        lodash \
        closure-compiler \
        uglify-js \
        grunt-contrib-uglify \
        grunt-modernizr \
        grunt-lodash \
        grunt-bower-task \
        grunt-shell \
        jison \
        jquery-builder \
        phantomjs \
        grunt-casperjs

(that is, listing all the packages in the `dependencies` and
`devDependencies` fields of package.json.

Once that error was overcome, I was getting this:

    Running "lodash:build" (lodash) task
    >> Error:
    >> Invalid argument passed: --debug
    >> For more information type: lodash --help
    Warning:
    Invalid argument passed: --debug
    For more information type: lodash --help Use --force to continue.

...which I fixed with the change to Gruntfile.js.

[1] Actually I do
    `npm install; PATH=$(pwd)/node_modules/.bin:$PATH grunt build`
    because I don't want to install grunt globally.